SQL Project: Music Store Analysis
DESCRIPTION
This project uses SQL to analyze the dataset of an online music store. The goal of the project is to answer a set of questions about the store's business performance and help in its growth by making better decisions.

INSTALLATION 
To run this project on your machine you need to install any SQL-supported DBMS then follow the steps below:

Create a database.
Create tables using the schema diagram provided below.
Import csv files provided in the dataset folder.
Note: If you are using PostgreSQL then restore the backup file of the database and save yourself some time.